server/backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2-binary
gunicorn
openai>=1.0.0
python-dotenv
httpx
google-generativeai
django-cors-headers
rest_framework_simplejwt
djangorestframework-simplejwt
